feat(sidebar): refactor sidebar project handling and introduce utility functions
